Fixed problem with simulation
To get access to the list of UUIDs to be queried for the simulation, a
pre-run query is executed, typically "select pkid from generic1 limit
10000;" However, this turns out to be a really bad query since the data
is by default sorted by the yb_hash_code. Hence the first many rows will
be in tablet 1, then spill over to tablet 2 and so on until the 10000
limit is hit. This is unlikely to be all tservers and might result in a
load skew issue with some servers seeing no traffic at the tserver
layer. Added a modulo in the code to try to fix this.